# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: North API
  version: 1.0.0
  description: |
    JSON API for North's native clients. The iOS app generates its client from
    this file (swift-openapi-generator), so this file is the contract: a route
    under /api/v1 that is not described here fails cmd/web's contract test.

    Authentication, by route group:
    - `/auth/*` sign-in routes take no credentials and return a session token.
    - Everything else takes that session token as `Authorization: Bearer`.
    - `/capture/*` takes an `nk_` connection token instead, issued from
      Settings → Connections for agents and shortcuts.

    A browser session cookie authenticates none of these.
servers:
  - url: https://kheprios.com/api/v1
    description: Production
  - url: http://localhost:8090/api/v1
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /supplements:
    get:
      operationId: getSupplements
      summary: Today's supplements and the preset list
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplementsToday'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: logSupplement
      summary: Log a supplement taken; answers with today
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplementInput'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplementsToday'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /supplements/{entryID}:
    parameters:
      - name: entryID
        in: path
        required: true
        description: The entry to undo.
        schema:
          type: string
    delete:
      operationId: undoSupplement
      summary: Undo a supplement
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplementsToday'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /soreness:
    get:
      operationId: getSoreness
      summary: Today's sore regions and the list of regions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SorenessToday'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /soreness/{region}:
    parameters:
      - name: region
        in: path
        required: true
        description: One of the regions GET /soreness lists.
        schema:
          type: string
    put:
      operationId: setSoreness
      summary: Set today's soreness in one region
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SorenessInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SorenessToday'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: clearSoreness
      summary: Clear today's soreness in one region
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SorenessToday'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /trackers:
    get:
      operationId: listTrackers
      summary: Months-since trackers (dentist, haircut)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createTracker
      summary: Start tracking something
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackerInput'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /trackers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: The tracker.
        schema:
          type: string
    delete:
      operationId: deleteTracker
      summary: Stop tracking
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /trackers/{id}/done:
    parameters:
      - name: id
        in: path
        required: true
        description: The tracker.
        schema:
          type: string
    post:
      operationId: markTrackerDone
      summary: Reset a tracker to today
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackerList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token returned by the /auth routes.
    connectionToken:
      type: http
      scheme: bearer
      description: An `nk_` connection token from Settings → Connections.
...
  responses:
...
    NotFound:
      description: No such resource for this account
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    Unauthorized:
      description: Missing or invalid bearer token
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    InternalError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
...
    SupplementInput:
      type: object
      additionalProperties: false
      properties:
        preset:
          type: string
        name:
          type: string
        count:
          type: integer
        nutrients:
          type: array
          items:
            type: string
    SupplementEntry:
      type: object
      required: [id, name, count, nutrients, loggedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        count:
          type: integer
        nutrients:
          type: array
          items:
            type: string
        loggedAt:
          type: string
          format: date-time
    SupplementPreset:
      type: object
      required: [key, name, nutrients]
      properties:
        key:
          type: string
        name:
          type: string
        nutrients:
          type: array
          items:
            type: string
    SupplementsToday:
      type: object
      required: [entries, presets, nutrients]
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/SupplementEntry'
        presets:
          type: array
          items:
            $ref: '#/components/schemas/SupplementPreset'
        nutrients:
          type: array
          items:
            type: string
...
    SorenessInput:
      type: object
      required: [severity]
      additionalProperties: false
      properties:
        severity:
          type: integer
          description: 1 stiff, 2 sore, 3 painful.
        note:
          type: string
    SorenessEntry:
      type: object
      required: [region, severity]
      properties:
        region:
          type: string
        severity:
          type: integer
        note:
          type: string
    SorenessToday:
      type: object
      required: [entries, regions]
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/SorenessEntry'
        regions:
          type: array
          items:
            type: string
    TrackerInput:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
        lastDoneOn:
          type: string
          format: date
        intervalMonths:
          type: integer
    Tracker:
      type: object
      required: [id, name, lastDoneOn, monthsSince, due]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        lastDoneOn:
          type: string
          format: date
        intervalMonths:
          type: integer
        monthsSince:
          type: integer
        due:
          type: boolean
    TrackerList:
      type: object
      required: [trackers]
      properties:
        trackers:
          type: array
          items:
            $ref: '#/components/schemas/Tracker'
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required: [message]
      properties:
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
...